+++ lib/src/code_generators/v2/swagger_requests_generator_v2.dart
@@ -1,5 +1,6 @@
 import 'dart:convert';
 import 'package:swagger_dart_code_generator/src/code_generators/swagger_requests_generator.dart';
+import 'package:swagger_dart_code_generator/src/code_generators/v2/swagger_models_generator_v2.dart';
 import 'package:swagger_dart_code_generator/src/models/generator_options.dart';
 import 'package:swagger_dart_code_generator/src/swagger_models/requests/swagger_request.dart';
 import 'package:swagger_dart_code_generator/src/swagger_models/requests/swagger_request_parameter.dart';
@@ -26,7 +27,10 @@ class SwaggerRequestsGeneratorV2 extends SwaggerRequestsGenerator {
 
     final definitions = map['definitions'] as Map<String, dynamic>;
 
+    final allEnumNames =
+        SwaggerModelsGeneratorV2().getAllEnumNames(definitions, code);
+
     return getFileContent(swaggerRoot, className, fileName, options,
-        definitions != null && definitions.keys.isNotEmpty);
+        definitions != null && definitions.keys.isNotEmpty, allEnumNames);
   }
 }